docs: update license and contributing information
After thinking it over for a while, I have decided to switch Tertiary
Text over to a less-restrictive license than the GPL. I have received
permission to do this from the authors of previous patches to Tertiary
Text.

As interesting as it would be to see a world where all code is forced to
be open-sourced, it doesn't seem to be realistic, and moreover it
doesn't seem to be the right thing to do. I'd like to get Tertiary Text
to be used by as many people who want to use it as possible, which means
not restricting developers to include it based on the status of their
own application's license.

This blog post: http://www.linuxjournal.com/article/5935 summarizes
pretty well how I feel about the issue right now. While I can see the
argument for how GPL is "more free," I personally don't want to include
this restriction at this time for developers who wish to use Tertiary
Text.